feat(web-ui): add providers page and restructure navigation
Create new Providers page combining provider connections (moved from
Profile) with model registry (moved from Models). Strip Profile page
to credentials-only with security section for password users. Update
navigation so all users see profile + providers, admins additionally
see config, guardrails, mcp servers, admin. Closes #61.
