+++ frontend/src/components/Layout.tsx
@@ -31,7 +31,7 @@ export function Layout() {
     if (location.pathname.includes('/config')) return 'configuration'
     if (location.pathname.includes('/guardrails')) return 'guardrails'
     if (location.pathname.includes('/mcp')) return 'mcp servers'
-    if (location.pathname.includes('/models')) return 'models'
+    if (location.pathname.includes('/providers')) return 'providers'
     if (location.pathname.includes('/profile')) return 'profile'
     if (location.pathname.includes('/admin')) return 'administration'
     return 'dashboard'
