+++ frontend/src/components/Sidebar.tsx
@@ -32,6 +32,9 @@ export function Sidebar({ open, onClose }: SidebarProps) {
         <NavLink to="/profile" className={({ isActive }) => `nav-link${isActive ? ' active' : ''}`} onClick={onClose}>
           <span className="nav-cursor">{'>'}</span> profile
         </NavLink>
+        <NavLink to="/providers" className={({ isActive }) => `nav-link${isActive ? ' active' : ''}`} onClick={onClose}>
+          <span className="nav-cursor">{'>'}</span> providers
+        </NavLink>
         {user.role === 'admin' && (
           <>
             <NavLink to="/config" className={({ isActive }) => `nav-link${isActive ? ' active' : ''}`} onClick={onClose}>
@@ -42,9 +45,6 @@ export function Sidebar({ open, onClose }: SidebarProps) {
             </NavLink>
             <NavLink to="/mcp" className={({ isActive }) => `nav-link${isActive ? ' active' : ''}`} onClick={onClose}>
               <span className="nav-cursor">{'>'}</span> mcp servers
-            </NavLink>
-            <NavLink to="/models" className={({ isActive }) => `nav-link${isActive ? ' active' : ''}`} onClick={onClose}>
-              <span className="nav-cursor">{'>'}</span> models
             </NavLink>
             <NavLink to="/admin" className={({ isActive }) => `nav-link${isActive ? ' active' : ''}`} onClick={onClose}>
               <span className="nav-cursor">{'>'}</span> admin
